cleanup return values and tri macro calls
Use arena's where possible to simplify deallocation.

Only downside to this is LuaRefs getting potentially leaked so we should
probably introduce an alternative clone logic for edge cases.

In practice all documented cases are handled but undocumented LuaRefs
will be leaked where cloning Object's, Array's and Dict's

+++ src/nvim_types/core/call_site.rs
@@ -1,6 +1,8 @@
+use super::Integer;
+
 #[repr(transparent)]
 #[derive(Clone, Copy, Debug, PartialEq, Eq)]
-pub struct Channel(i64);
+pub struct Channel(Integer);
 
 impl Channel {
     // Simplified version of src/nvim/api/private/defs.h 0.10.0 l:42
